requirements.txt
click>=8.1.3
discord.py>=2.0.0
snakecore @ git+https://github.com/pygame-community/snakecore.git@v0.1.1#egg=snakecore-0.1.1
uvloop; sys_platform != "win32" # https://github.com/MagicStack/uvloop/issues/14
